Fixes for windows
Also fixes some warning

Important note: on windows, we only generate a static library for now.
With a custom build of SFML it should be working with dynamic linking but
with the SFML-2.1 distributed by sfml-dev.org that's not the case.

+++ src/OVRWrapper.cpp
@@ -93,7 +93,7 @@ OVRWrapper::OVRWrapper(float width, float height, bool enable_scaling)
     {
         sfusion = new OVR::SensorFusion(sensor); 
     }
-    sConfig.SetFullViewport(OVR::Util::Render::Viewport(0,0, width, height));
+    sConfig.SetFullViewport(OVR::Util::Render::Viewport(0,0, (int)width, (int)height));
     sConfig.SetStereoMode(OVR::Util::Render::Stereo_LeftRight_Multipass);
     sConfig.SetDistortionFitPointVP(-1.0f, 0.0f);
     renderScale = enable_scaling ? sConfig.GetDistortionScale() : 1;
